+++ src/contracts/Money.sol
@@ -3,29 +3,29 @@ pragma solidity >= 0.8.7 <= 0.9;
 //pragma solidity >=0.4.22 <0.9.0; 
 
 contract Money {
-    uint balance = 0;
+    uint256 balance = 0;
 
 
-    modifier validateDeposit(uint amount) {
+    modifier validateDeposit(uint256 amount) {
         require(amount > 0, "Invalid amount.");
         _;
     }
 
-    modifier validateWithdrawal(uint amount) {
+    modifier validateWithdrawal(uint256 amount) {
         require(amount > 0, "Invalid amount.");
         require(amount <= balance, "Insufficient balance.");
         _;
     }
 
-    event depositCompplete(uint amount);
-    event withdrawalComplete(uint amount);
+    event depositCompplete(uint256 amount);
+    event withdrawalComplete(uint256 amount);
 
-    function Deposit(uint amount) external validateDeposit(amount) {
+    function Deposit(uint256 amount) external validateDeposit(amount) {
         balance += amount;
         emit depositCompplete(amount);
     }
 
-    function Withdraw(uint amount) external validateWithdrawal(amount) {
+    function Withdraw(uint256 amount) external validateWithdrawal(amount) {
         balance -= amount;
         emit withdrawalComplete(amount);
     }
